Allow metadata to be retrieved from GetCapabilities
Reorganise metadata code and make it a little more resilient
