+++ GIFrameworkMaps.Web/Scripts/Interfaces/OGCMetadata/LayerResource.ts
@@ -13,4 +13,5 @@ export interface LayerResource {
     version: string;
     proxyMetaRequests: boolean;
     proxyMapRequests: boolean;
+    keywords?: string[];
 }
